Fixed inconsistent default color in `vp.write_svg()`.
The default color (from the default color scheme) for a layer would depend on whether previous layer would have their color specified on not. This is an inconsistent behaviour in itself, and it is inconsistent with the viewer's behaviour.

Also, the default color scheme as been moved to vpype/metadata.py.