Feature: property substitution and block overhaul
The input provided to most commands' arguments and options may now contain substitution patterns which will be replaced by the corresponding property value.

This created the need for deferred evaluation of click-provided arguments and options. The type classes now return subclass of `_DeferredEvaluator` which are detected by the command decorators (`@generator`, etc.) and replaced by their evaluated value before they reach the command function.

In turn, implementing the above lead to the overhaul of the block processor sub-system. Block processors are now implemented as functions and the nested command have access to the current layer structure.

* Added property substitution function to `vpype_cli.State`
* Added mechanism to perform deferred evaluation of Click argument/option value
* Added `TextType` class to use as Click type instead of `str` so property substitution is performed.
* Updated types to perform property substitution
* `grid` now iterates along lines first
* Document.clone() now has an option to keep the layer structure/metadata (without line data)
* fixed an issue where metadata dict could be shared across Doc/LC, leading to potential bugs.
* fixed a bug with `random` when using non-square area
* Various docstring fixes and improvements
* `assert_image_similarity` now skips the test when the reference image is saved