Fixed missing plug-ins from `vpype --help` string (#444)
When importing plug-ins in the style of `click-plugin` (command decorator, so plug-ins are loaded during the loading of the top-level command itself), plug-ins may not import from `vpype_cli` (since it isn't fully loaded). To address that, since 1.9, loading plug-ins is deferred to when `cli` is actually executed. As a result, the Click's default behaviour for handling `--help` (i.e. print and exit *before* even executing `cli`) is unable to list the plug-ins. This commit addresses this by manually handling the top-level `--help` parameter (*after* plug-ins are loaded).

Also fixed multiple plug-in loading when using `vpype_cli.execute()` multiple times

Fixes #432