Improved handling of SVG with missing width and height.
When width and height are missing from the <svg> tag, or expressed as percent, we now let svgelement use viewBox (if available) or revert to 1000x1000. This happens by default now with `read_svg()` (and friends). This behaviour can be overridden by passing values for `default_width` and `default_height`.

The `read` command now follows this pattern as well.

Fixes #441