Improved handling of pen width and opacity in viewer (#448)
This PR improves the viewer so it honours the pen width and opacity (`vp_color`'s alpha channel) when present. The UI is updated to reflect that the choosable pen width/opacity are now default value, used when they are undefined. It also adds options to override the pen width/opacity from the properties and use the default value instead.

- The viewer now honors pen width and opacity
- The corresponding menu items have now been renamed "Default pen opacity" and "Default pen width"
- Added an "Override" menu item to both menu to override pen width/opacity from properties
- Fixed rendering inconsistence (integer coordinate for vert/horiz line) with the rulers and `test_text_block_render()`, which fixed the failing tests on M1 Macs
- Added tests for pen opacity/width override
- API (breaking change): engine renamed `pen_width`/`pen_opacity` property to `default_pen_width`/`default_pen_opacity`
- API (breaking change): renamed parameters of `vpype_viewer.render_image()`

Fixes #377 
Fixes #375